requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
python-dotenv==1.0.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.16
langchain-openai==0.0.8
langgraph==0.0.40
openai==1.102.0

# OCR and document processing
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
paddlepaddle==2.5.2
paddleocr==2.7.0

# Cloud OCR services (optional)
boto3==1.34.0
azure-ai-formrecognizer==3.3.0

# Data processing
pandas==2.3.2
numpy==1.26.4
pydantic==2.11.7

# Logging and monitoring
structlog==23.2.0

# Utilities
requests==2.32.5
Pillow==10.4.0

# Database (desabilitado temporariamente)
# sqlalchemy==2.0.43
# psycopg2-binary==2.9.10
# alembic==1.16.4
